refactor: 메서드, 클래스 이름 변경
